Move link table to seperate page
