+++ apps/dashboard/src/routes/dashboard/link/LinkDetailsModal.svelte
